renamed loadConfig to applyConfig
