+++ addon/utils/apply-config.js
