+++ tests/unit/utils/apply-config-test.js
@@ -2,11 +2,11 @@
 import { it } from 'ember-mocha';
 import { describe, beforeEach } from 'mocha';
 import { expect } from 'chai';
-import loadConfig from 'ember-simple-auth/utils/load-config';
+import applyConfig from 'ember-simple-auth/utils/apply-config';
 
-describe('loadConfig', () => {
+describe('applyConfig', () => {
   it('returns a function', () => {
-    expect(loadConfig()).to.be.a('function');
+    expect(applyConfig()).to.be.a('function');
   });
 
   describe('the application function', () => {
@@ -15,32 +15,32 @@ describe('loadConfig', () => {
     beforeEach(() => {
       object = {
         setting:    null,
-        loadConfig: loadConfig({ setting: 'defaultValue' })
+        applyConfig: applyConfig({ setting: 'defaultValue' })
       };
     });
 
     it('applies a value if present on the config object', () => {
-      object.loadConfig({ setting: 'configValue' });
+      object.applyConfig({ setting: 'configValue' });
 
       expect(object.setting).to.eq('configValue');
     });
 
     it('applies the default value when a setting is not present on the config object', () => {
-      object.loadConfig({});
+      object.applyConfig({});
 
       expect(object.setting).to.eq('defaultValue');
     });
 
     it('ignores settings that are not properties of the configured object', () => {
-      object.loadConfig({ unknown: 'configValue' });
+      object.applyConfig({ unknown: 'configValue' });
 
       expect(object).to.not.haveOwnProperty('unknown');
     });
 
     it('does not change methods of the configured object', () => {
-      object.loadConfig({ loadConfig: 'configValue' });
+      object.applyConfig({ applyConfig: 'configValue' });
 
-      expect(object.loadConfig).to.not.eq('configValue');
+      expect(object.applyConfig).to.not.eq('configValue');
     });
   });
 });
